przejscie na bigDecimal, poprawienie mapowan, porzadki w pomie

+++ src/main/java/pl/nutrivia/domain/HasCalories.java
@@ -1,6 +1,8 @@
 package pl.nutrivia.domain;
 
+import java.math.BigDecimal;
+
 @FunctionalInterface
 public interface HasCalories {
-    double getCalories();
+    BigDecimal getCalories();
 }
